chore(nextjs): Improve observability on keyless failures
