+++ packages/nextjs/src/app-router/server/logger.ts
@@ -1,0 +1,4 @@
+import { createDebugLogger } from '../../utils/debugLogger';
+import { logFormatter } from '../../utils/logFormatter';
+
+export const ClerkProviderDebugLogger = createDebugLogger('<ClerkProvider/>', logFormatter)();
